Improve hotspot_layout to allow for internal verts
Previously the hotspot layout tool would only work on rectangles with no internal vertices. Now this code has been updated so that you can select a UV shell with internal vertices and it will properly lay out to a hotspot, keeping the relative distance of internal verts correctly. Big improvement with this one.